16-authentication/requirements.txt
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
authlib==1.2.1
httpx==0.25.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
itsdangerous==2.1.2
pyotp==2.9.0
qrcode==7.4.2
redis==5.0.1
aioredis==2.0.1
